telegram-bot/requirements.txt
paho-mqtt
python-telegram-bot
